requirements.txt
pandas
numpy
scikit-learn
scipy
pyyaml
requests
tqdm
streamlit
pytest
poke-env
